Refactor the snippet listing.
Add a formatCfg type holding details of how a snippet should be displayed;
the parts of the snippet and whether to show the names of the parts.

Change snippet.String to use the formatCfg type

Add a ListCfg type that holds configuration details controlling the way that
a snippet list is displayed. It also has an embedded formatCfg so that it
controls the formatting of individual snippets as well. With this it is now
possible to control which snippets are to be shown and which parts of the
snippet.

The semantic snippet comments are now case insensitive and allow for some
alias values. So for instance you can label the documentation comments with
any one of 'doc', 'docs', 'Docs', 'NoTeS' etc.